+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Dates in this file are in [Holocene Calendar] because it is amazing, logical, an
 
 ## [Unreleased]
 
+## [v0.2.1] 12024-10-22
+
 ### Added
   - Logs are now saved to a file in the state directory for troubleshooting
   - Material pipette: a checkbox to automatically check off held materials in the checklist
@@ -32,7 +34,8 @@ Dates in this file are in [Holocene Calendar] because it is amazing, logical, an
 ### Added
   - The first release
 
-[unreleased]: https://github.com/necauqua/noita-utility-box/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/necauqua/noita-utility-box/compare/v0.2.1...HEAD
+[v0.2.1]: https://github.com/necauqua/noita-utility-box/releases/tag/v0.2.1
 [v0.2.0]: https://github.com/necauqua/noita-utility-box/releases/tag/v0.2.0
 [v0.1.0]: https://github.com/necauqua/noita-utility-box/releases/tag/v0.1.0